take out individual health scripts, add common game manager
