--- Assets/Scripts/SquidMainHealthSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a831c08445942fd479117a52c751ecbe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
